+++ eventlog-live/CHANGELOG.md
@@ -7,6 +7,7 @@
 - **BREAKING**: Drop `Stream` and `supplier` machine.
 - **BREAKING**: Rename `GHC.Eventlog.Live.Machines` to `GHC.Eventlog.Live.Machine`.
 - **BREAKING**: Move core machines to their own module.
+- **BREAKING**: Use ANSI colors for logs, if supported.
 
 ### 0.1.0.1